refactor: dropping diesel + anyhow

+++ libs/rs2-communication-protocol/src/communication_structs/checkin.rs
@@ -94,9 +94,9 @@ pub struct CheckinResponse {
     /// The working hours of for the current day (unix timestamp), if any
     pub working_hours:    Option<Vec<Option<i64>>>,
     /// The agent polling interval in milliseconds
-    pub polling_interval: i64,
+    pub polling_interval: u128,
     /// The agent polling jitter in milliseconds
-    pub polling_jitter:   i64,
+    pub polling_jitter:   u128,
 }
 
 #[cfg(test)]
